feat: auth 라우트 핸들러 설정 (#63)
