+++ src/auth.ts
@@ -1,0 +1,27 @@
+import NextAuth from 'next-auth';
+import Kakao from 'next-auth/providers/kakao';
+
+export const { handlers, auth, signIn, signOut } = NextAuth({
+  providers: [
+    Kakao({
+      clientId: process.env.AUTH_KAKAO_ID!,
+      clientSecret: process.env.AUTH_KAKAO_SECRET!,
+    }),
+  ],
+  session: { strategy: 'jwt' },
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) {
+        token.userId = user.id ?? token.sub;
+      }
+      return token;
+    },
+    async session({ session, token }) {
+      session.user = {
+        ...(session.user ?? {}),
+        id: (token.userId ?? token.sub ?? '') as string,
+      };
+      return session;
+    },
+  },
+});
